Add new Certificate Authority
Old certificate authority will expire in July 2017. Adding the
certificate of the authority that will replace it. According to
https://community.openvpn.net/openvpn/wiki/Using_Certificate_Chains
(see mention of stacked certificates), it is allowed to simply
include all chains in the ca part of the configuration. The old
pair remains and will remain at least until June 2017.